+++ semver-project-gradle-plugin/main/kotlin/com/javiersc/semver/project/gradle/plugin/LazyVersion.kt
@@ -6,6 +6,10 @@ public class LazyVersion(internal val version: Provider<String>) {
 
     private var cachedVersion: String? = null
 
+    public fun map(transform: (String) -> String) {
+        cachedVersion = transform(version.get())
+    }
+
     override fun toString(): String {
         if (cachedVersion == null) cachedVersion = version.get()
         cachedVersion ?: version.get()